+++ tests/test_plaintext_parser.py
@@ -66,3 +66,97 @@ MLS #963378
         assert listings[0].price == 1500000
         assert listings[0].bedrooms == 4
         assert listings[0].mls_id == "123456"
+
+    def test_extracts_redfin_url(self):
+        text = """
+11 Jennifer Lane
+Rye Brook, New York 10573
+$1,295,000
+4 bd, 3 ba, 2,437 sqft
+MLS #964038
+https://www.redfin.com/NY/Rye-Brook/11-Jennifer-Ln-10573/home/22583423
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 1
+        assert listings[0].listing_url == "https://www.redfin.com/NY/Rye-Brook/11-Jennifer-Ln-10573/home/22583423"
+
+    def test_extracts_onekeymls_url(self):
+        text = """
+$1,400,000 - 3 bd, 3 ba
+MLS #963378
+https://www.onekeymls.com/listing/963378
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 1
+        assert listings[0].listing_url == "https://www.onekeymls.com/listing/963378"
+
+    def test_strips_trailing_punctuation_from_url(self):
+        text = """
+$1,295,000 - 4 bd
+MLS #964038
+Check it out: https://www.redfin.com/NY/Rye-Brook/11-Jennifer-Ln/home/123.
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 1
+        assert listings[0].listing_url == "https://www.redfin.com/NY/Rye-Brook/11-Jennifer-Ln/home/123"
+
+    def test_no_url_when_absent(self):
+        text = """
+$1,295,000
+4 bd, 3 ba, 2,437 sqft
+MLS #964038
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 1
+        assert listings[0].listing_url is None
+
+    def test_filters_redfin_tour_urls(self):
+        """Tour/checkout URLs should NOT be extracted as listing URLs."""
+        text = """
+$1,200,000 - 4 bd, 2.5 ba, 3,034 sqft
+https://www.redfin.com/tours/checkout/times?listingId=212530348
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 1
+        assert listings[0].listing_url is None
+
+    def test_inline_address_format(self):
+        """Redfin-style inline address: '31 Lalli Dr, Katonah, NY 10536'."""
+        text = """
+$1,200,000 4 Beds 2.5 Baths 3,034 sqft
+31 Lalli Dr, Katonah, NY 10536
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 1
+        assert listings[0].address == "31 Lalli Dr"
+        assert listings[0].town == "Katonah"
+        assert listings[0].state == "NY"
+        assert listings[0].zip_code == "10536"
+
+    def test_inline_address_multi_word_town(self):
+        """Inline address with multi-word town name."""
+        text = """
+$1,695,000 5 Beds 3.5 Baths 3,274 sqft
+7 Dickson Ln, Bedford Corners, NY 10549
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 1
+        assert listings[0].address == "7 Dickson Ln"
+        assert listings[0].town == "Bedford Corners"
+        assert listings[0].zip_code == "10549"
+
+    def test_redfin_tour_email_two_listings(self):
+        """Tour confirmation email with two inline-address listings."""
+        text = """
+$1,200,000 4 Beds 2.5 Baths 3,034 Sq. Ft.
+31 Lalli Dr, Katonah, NY 10536
+
+$1,695,000 5 Beds 3.5 Baths 3,274 Sq. Ft.
+7 Dickson Ln, Bedford Corners, NY 10549
+        """
+        listings = self.parser.parse(None, text)
+        assert len(listings) == 2
+        assert listings[0].address == "31 Lalli Dr"
+        assert listings[0].price == 1200000
+        assert listings[1].address == "7 Dickson Ln"
+        assert listings[1].price == 1695000